Made Test Class for Calendar
Only at 95% coverage, will do more later, but can only do so much cause of the private method that checks for legal Events. Look at it later.

AS BEFORE:

Start on front end to make initial Calendar and test front end to back end. Will do comprehensive testing to make sure there are no bugs in the initial code and will fix any code in backend. After Knowing it all works, will start the scheduler portion (class) of the project, which will take a lot of work and Algo.

Before I work On the front-end (no design and just buttons) make all the method comments and file headers to ensure clarity in code and to show all work is mine and can't be used unless asked.