Commented/JavaDoc all of Calendar methods and class header
Use JavaDoc to make it easier to read when coding and reviewing the code. If any other methods were to be refactored or added, will update/add the JavaDoc for it, to make sure accurate information is being laid out.

Will do this to the rest of the Classes one at a time for consistency throughout the code. Don't want to rush it and should do it when I don't have enough time to actually code, but still have enough time to do some work.

NOTES: I think the view classes are bugged out from the refactor, might need to save->delete->make new classes->paste to use them again. (I don't know why it stopped working, will do reaserch to see if I can fix it).