Connected the Login/Signup with Main Menu/Settings/BOD
Not only does everything Connect, but it now has the proper error messages that pop up.

I need to make the settings More usable (make it so that user can pick color for the calendar and such)

Need to make BOD actually work, right now, it is just a picture that the user can look at. Need to add its functionality.

Next I need to start on the actual the frontend of the calendar, will most likely do it last since it will be the hardest to do and use the Observer pattern. Need to do some research on best ways to implement it to get some inspiration.